refactor tests that use FakeFS
