+++ test/options_parsing_test.rb
@@ -2,6 +2,7 @@ require 'test/helper'
 
 class OptionParsingTest < Test::Unit::TestCase 
   include Heroku::Kensa
+  include FsMock
 
   def options_for_cmd(string)
     client = Client.new(string.split)
